% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mkEMRun}
\alias{mkEMRun}
\title{Creates an object of type EMRun}
\usage{
mkEMRun()
}
\value{
A new object of type EMRun, with a matrix mu, covariance matrix sigma, a matrix with weights for each sample point (weights), a probability for each cluster (clusterProbs), a log-likelihhod (logL) and a list of data
}
\description{
Creates an object of type EMRun
}
