% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mkFlowDataObject}
\alias{mkFlowDataObject}
\title{Creates a flow data object.}
\usage{
mkFlowDataObject(frame, xChannel, yChannel)
}
\arguments{
\item{frame}{Fill me}

\item{xChannel}{...}

\item{yChannel}{...}
}
\description{
Fill me!
}
