% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mkFlowDataObject}
\alias{mkFlowDataObject}
\title{Creates a flow data object from a frame.}
\usage{
mkFlowDataObject(frame, xChannel, yChannel)
}
\arguments{
\item{frame}{FCS data frame.}

\item{xChannel}{"x"-channel to extract.}

\item{yChannel}{"y"-channel to extract.}
}
\value{
flow data object
}
\description{
This frame is typically the result of a call to read.FCS().
}
